requirements.txt
selenium >= 4.41.0
python-dotenv >= 1.2.2
pandas >= 3.0.1